+++ app/screens/Creature/Styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+const Styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        marginTop: 20,
+    },
+    title: {
+        fontSize: 23,
+        marginBottom: 20,
+    },
+    description: {
+        fontSize: 18,
+        marginTop: 20,
+    }
+});
+
+export default Styles;