+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #v0.2.0
 
 - Change the home screen.
+- Add the creature screen.
 
 #v0.1.0